+++ examples/print-oracle-version.rs
@@ -21,7 +21,7 @@ fn main() {
     let client_ver = client_version().unwrap();
     println!("Oracle Client Version: {}", client_ver);
 
-    let conn = Connection::connect("scott", "tiger", "", &[]).unwrap();
+    let conn = Connection::connect("scott", "tiger", "").unwrap();
     let (server_ver, banner) = conn.server_version().unwrap();
     println!("Oracle Server Version: {}", server_ver);
     println!("--- Server Version Banner ---");