+++ src/connection.rs
@@ -87,51 +87,111 @@ pub enum ShutdownMode {
     Final,
 }
 
-/// Parameters to create Connection passed to [Connection::connect][].
+/// [Administrative privilege](https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/configuring-privilege-and-role-authorization.html#GUID-633842B8-4B19-4F96-A757-783BF62825A7)
 ///
-/// [Connection::connect]: struct.Connection.html#method.connect
+/// See [Connector.privilege](struct.Connector.html#method.privilege).
+#[derive(Debug, Copy, Clone, PartialEq)]
+pub enum Privilege {
+    /// Connects as [SYSDBA](https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/configuring-privilege-and-role-authorization.html#GUID-BD5D39D1-DBFF-400A-8645-355F8FB9CD31).
+    ///
+    Sysdba,
+
+    /// Connects as [SYSOPER](https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/configuring-privilege-and-role-authorization.html#GUID-BD5D39D1-DBFF-400A-8645-355F8FB9CD31).
+    Sysoper,
+
+    /// Connects as [SYSASM](https://docs.oracle.com/en/database/oracle/oracle-database/19/ostmg/authenticate-access-asm-instance.html#GUID-7396FD18-628B-4026-AA55-79C6D6205EAE) (Oracle 12c or later)
+    Sysasm,
+
+    /// Connects as [SYSBACKUP](https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/configuring-privilege-and-role-authorization.html#GUID-BF12E37F-4606-42BB-B8B6-4CDC5A870EE7)
+    Sysbackup,
+
+    /// Connects as [SYSDG](https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/configuring-privilege-and-role-authorization.html#GUID-5798F976-85B2-4973-92F7-DB3F6BC9D497) (Oracle 12c or later)
+    Sysdg,
+
+    /// Connects as [SYSKM](https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/configuring-privilege-and-role-authorization.html#GUID-573B5831-E106-4D8C-9101-CF9C1B74A39C) (Oracle 12c or later)
+    Syskm,
+
+    /// Connects as [SYSRAC](https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/configuring-privilege-and-role-authorization.html#GUID-69D0614C-D24E-4EC1-958A-79D7CCA3FA3A) (Oracle 12c R2 or later)
+    Sysrac,
+}
+
+#[derive(Debug, Copy, Clone, PartialEq)]
+/// [Session Purity](https://docs.oracle.com/en/database/oracle/oracle-database/19/adfns/performance-and-scalability.html#GUID-12410EEC-FE79-42E2-8F6B-EAA9EDA59665)
+pub enum Purity {
+    /// Must use a new session
+    New,
+    /// Reuse a pooled session
+    Self_,
+}
+
+/// Builder data type to create Connection.
+///
+/// When a connection can be established only with username, password
+/// and connect string, use [Connection.connect] instead.
+///
+/// [Connection.connect]: struct.Connection.html#method.connect
 #[derive(Debug, Clone, PartialEq)]
-pub enum ConnParam {
-    /// Connects as [SYSDBA](https://docs.oracle.com/database/122/ADMQS/administering-user-accounts-and-security.htm#GUID-2033E766-8FE6-4FBA-97E0-2607B083FA2C)
+pub struct Connector {
+    username: String,
+    password: String,
+    connect_string: String,
+    privilege: Option<Privilege>,
+    external_auth: bool,
+    prelim_auth: bool,
+    new_password: String,
+    purity: Option<Purity>,
+    connection_class: String,
+    app_context: Vec<(String, String, String)>,
+    tag: String,
+    match_any_tag: bool,
+    events: bool,
+    edition: String,
+    driver_name: String,
+}
+
+impl Connector {
+    /// Create a connector
+    pub fn new<U, P, C>(username: U, password: P, connect_string: C) -> Connector
+    where
+        U: Into<String>,
+        P: Into<String>,
+        C: Into<String>,
+    {
+        Connector {
+            username: username.into(),
+            password: password.into(),
+            connect_string: connect_string.into(),
+            privilege: None,
+            external_auth: false,
+            prelim_auth: false,
+            new_password: "".into(),
+            purity: None,
+            connection_class: "".into(),
+            app_context: vec![],
+            tag: "".into(),
+            match_any_tag: false,
+            events: false,
+            edition: "".into(),
+            driver_name: "".into(),
+        }
+    }
+
+    /// Set [administrative privilege](https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/configuring-privilege-and-role-authorization.html#GUID-633842B8-4B19-4F96-A757-783BF62825A7).
     ///
     /// # Examples
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// let conn = Connection::connect("system", "manager", "", &[ConnParam::Sysdba])?;
+    /// // connect system/manager as sysdba
+    /// let conn = Connector::new("system", "manager", "")
+    ///     .privilege(Privilege::Sysdba)
+    ///     .connect()?;
     /// # Ok(())} fn main() { try_main().unwrap(); }
     /// ```
-    Sysdba,
-
-    /// Connects as [SYSOPER](https://docs.oracle.com/database/122/ADMQS/administering-user-accounts-and-security.htm#GUID-2033E766-8FE6-4FBA-97E0-2607B083FA2C)
-    ///
-    /// See [`Examples`](#examples) in the Sysdba variant.
-    Sysoper,
-
-    /// Connects as [SYSASM](https://docs.oracle.com/database/122/OSTMG/authenticate-access-asm-instance.htm#OSTMG02600) (Oracle 12c or later)
-    ///
-    /// See [`Examples`](#examples) in the Sysdba variant.
-    Sysasm,
-
-    /// Connects as [SYSBACKUP](https://docs.oracle.com/database/122/DBSEG/configuring-privilege-and-role-authorization.htm#DBSEG785) (Oracle 12c or later)
-    ///
-    /// See [`Examples`](#examples) in the Sysdba variant.
-    Sysbackup,
-
-    /// Connects as [SYSDG](https://docs.oracle.com/database/122/DBSEG/configuring-privilege-and-role-authorization.htm#GUID-5798F976-85B2-4973-92F7-DB3F6BC9D497) (Oracle 12c or later)
-    ///
-    /// See [`Examples`](#examples) in the Sysdba variant.
-    Sysdg,
-
-    /// Connects as [SYSKM](https://docs.oracle.com/database/122/DBSEG/configuring-privilege-and-role-authorization.htm#GUID-573B5831-E106-4D8C-9101-CF9C1B74A39C) (Oracle 12c or later)
-    ///
-    /// See [`Examples`](#examples) in the Sysdba variant.
-    Syskm,
-
-    /// Connects as [SYSRAC](https://docs.oracle.com/database/122/DBSEG/configuring-privilege-and-role-authorization.htm#DBSEG-GUID-69D0614C-D24E-4EC1-958A-79D7CCA3FA3A) (Oracle 12c R2 or later)
-    ///
-    /// See [`Examples`](#examples) in the Sysdba variant.
-    Sysrac,
+    pub fn privilege(&mut self, privilege: Privilege) -> &mut Connector {
+        self.privilege = Some(privilege);
+        self
+    }
 
     /// Uses external authentication such as [OS authentication][].
     ///
@@ -139,17 +199,25 @@ pub enum ConnParam {
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// let conn = Connection::connect("", "", "", &[ConnParam::ExternalAuth])?;
+    /// let conn = Connector::new("", "", "")
+    ///     .external_auth(true)
+    ///     .connect()?;
     /// # Ok(())} fn main() { try_main().unwrap(); }
     /// ```
     ///
-    /// [OS authentication]: https://docs.oracle.com/en/database/oracle/oracle-database/12.2/dbseg/configuring-authentication.html#GUID-37BECE32-58D5-43BF-A098-97936D66968F
-    ExternalAuth,
+    /// [OS authentication]: https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/configuring-authentication.html#GUID-37BECE32-58D5-43BF-A098-97936D66968F
+    pub fn external_auth(&mut self, b: bool) -> &mut Connector {
+        self.external_auth = b;
+        self
+    }
 
     /// Sets prelim auth mode to connect to an idle instance.
     ///
     /// See [starting up a database](struct.Connection.html#method.startup_database).
-    PrelimAuth,
+    pub fn prelim_auth(&mut self, b: bool) -> &mut Connector {
+        self.prelim_auth = b;
+        self
+    }
 
     /// Sets new password during establishing a connection.
     ///
@@ -164,80 +232,113 @@ pub enum ConnParam {
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// let conn = match Connection::connect("scott", "tiger", "", &[]) {
+    /// let conn = match Connection::connect("scott", "tiger", "") {
     ///     Ok(conn) => conn,
     ///     Err(Error::OciError(ref dberr)) if dberr.code() == 28001 => {
     ///         // ORA-28001: the password has expired
-    ///         let params = [ConnParam::NewPassword("jaguar".into())];
-    ///         Connection::connect("scott", "tiger", "", &params)?
+    ///         Connector::new("scott", "tiger", "")
+    ///             .new_password("jaguar")
+    ///             .connect()?
     ///     },
     ///     Err(err) => panic!(err.to_string()),
     /// };
     /// # Ok(())} fn main() { try_main().unwrap(); }
     /// ```
-    NewPassword(String),
+    pub fn new_password<P>(&mut self, password: P) -> &mut Connector
+    where
+        P: Into<String>,
+    {
+        self.new_password = password.into();
+        self
+    }
 
-    /// Uses a new session in [DRCP][] pooled sessions.
+    /// Sets session purity specifying whether an application can reuse a pooled session (`Purity::Self_`) or must use a new session (`Purity::New`) from [DRCP][] pooled sessions.
     ///
-    /// See [here][] for more detail.
-    ///
-    /// [DRCP]: https://docs.oracle.com/en/database/oracle/oracle-database/12.2/adfns/performance-and-scalability.html#GUID-0B0DC596-17FF-4EEE-BAA1-9F428710C1AC
-    /// [here]: https://docs.oracle.com/en/database/oracle/oracle-database/12.2/adfns/performance-and-scalability.html#GUID-12410EEC-FE79-42E2-8F6B-EAA9EDA59665
-    PurityNew,
-
-    /// Reuses a pooled session in [DRCP][] pooled sessions.
-    ///
-    /// See [here][] for more detail.
-    ///
-    /// [DRCP]: https://docs.oracle.com/en/database/oracle/oracle-database/12.2/adfns/performance-and-scalability.html#GUID-0B0DC596-17FF-4EEE-BAA1-9F428710C1AC
-    /// [here]: https://docs.oracle.com/en/database/oracle/oracle-database/12.2/adfns/performance-and-scalability.html#GUID-12410EEC-FE79-42E2-8F6B-EAA9EDA59665
-    PuritySelf,
+    /// [DRCP]: https://docs.oracle.com/en/database/oracle/oracle-database/19/adfns/performance-and-scalability.html#GUID-015CA8C1-2386-4626-855D-CC546DDC1086
+    pub fn purity(&mut self, purity: Purity) -> &mut Connector {
+        self.purity = Some(purity);
+        self
+    }
 
     /// Sets a connection class to restrict sharing [DRCP][] pooled sessions.
     ///
     /// See [here][] for more detail.
     ///
-    /// [DRCP]: https://docs.oracle.com/en/database/oracle/oracle-database/12.2/adfns/performance-and-scalability.html#GUID-0B0DC596-17FF-4EEE-BAA1-9F428710C1AC
-    /// [here]: https://docs.oracle.com/en/database/oracle/oracle-database/12.2/adfns/performance-and-scalability.html#GUID-EC3DEE61-512C-4CBB-A431-91894D0E1E37
-    ConnectionClass(String),
+    /// [DRCP]: https://docs.oracle.com/en/database/oracle/oracle-database/19/adfns/performance-and-scalability.html#GUID-015CA8C1-2386-4626-855D-CC546DDC1086
+    /// [here]: https://docs.oracle.com/en/database/oracle/oracle-database/19/adfns/performance-and-scalability.html#GUID-EC3DEE61-512C-4CBB-A431-91894D0E1E37
+    pub fn connection_class<S>(&mut self, connection_class: S) -> &mut Connector
+    where
+        S: Into<String>,
+    {
+        self.connection_class = connection_class.into();
+        self
+    }
 
-    /// Sets an application context.
+    /// Appends an application context.
     ///
-    /// See [Oracle manual](https://docs.oracle.com/database/122/DBSEG/using-application-contexts-to-retrieve-user-information.htm#DBSEG165)
+    /// See [Oracle manual](https://docs.oracle.com/en/database/oracle/oracle-database/19/dbseg/using-application-contexts-to-retrieve-user-information.html#GUID-5841261E-988F-4A56-A2B4-71114AB3D51D)
     ///
     /// This is same with [DBMS_SESSION.SET_CONTEXT][] but this can set application contexts before a connection is established.
     ///
-    /// [DBMS_SESSION.SET_CONTEXT]: https://docs.oracle.com/database/122/ARPLS/DBMS_SESSION.htm#GUID-395C622C-ED79-44CC-9157-6A320934F2A9
+    /// [DBMS_SESSION.SET_CONTEXT]: https://docs.oracle.com/en/database/oracle/oracle-database/19/arpls/DBMS_SESSION.html#GUID-395C622C-ED79-44CC-9157-6A320934F2A9
     ///
     /// # Examples
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// let params = [
-    ///     ConnParam::AppContext("CLIENTCONTEXT".into(), "foo".into(), "bar".into()),
-    ///     ConnParam::AppContext("CLIENTCONTEXT".into(), "baz".into(), "qux".into()),
-    /// ];
-    /// let conn = Connection::connect("scott", "tiger", "", &params)?;
+    /// let conn = Connector::new("scott", "tiger", "")
+    ///               .app_context("CLIENTCONTEXT", "foo", "bar")
+    ///               .app_context("CLIENTCONTEXT", "baz", "qux")
+    ///               .connect()?;
+    /// let val = conn.query_row_as::<String>("select sys_context('CLIENTCONTEXT', 'foo') from dual", &[])?;
+    /// assert_eq!(val, "bar");
     /// let val = conn.query_row_as::<String>("select sys_context('CLIENTCONTEXT', 'baz') from dual", &[])?;
     /// assert_eq!(val, "qux");
     /// # Ok(())} fn main() { try_main().unwrap(); }
     /// ```
-    AppContext(String, String, String),
+    pub fn app_context<T1, T2, T3>(&mut self, namespace: T1, name: T2, value: T3) -> &mut Connector
+    where
+        T1: Into<String>,
+        T2: Into<String>,
+        T3: Into<String>,
+    {
+        self.app_context
+            .push((namespace.into(), name.into(), value.into()));
+        self
+    }
 
     /// Reserved for when connection pooling is supported.
-    Tag(String),
+    pub fn tag<S>(&mut self, tag: S) -> &mut Connector
+    where
+        S: Into<String>,
+    {
+        self.tag = tag.into();
+        self
+    }
 
     /// Reserved for when connection pooling is supported.
-    MatchAnyTag,
+    pub fn match_any_tag(&mut self, b: bool) -> &mut Connector {
+        self.match_any_tag = b;
+        self
+    }
 
     /// Reserved for when advanced queuing (AQ) or continuous query
     /// notification (CQN) is supported.
-    Events,
+    pub fn events(&mut self, b: bool) -> &mut Connector {
+        self.events = b;
+        self
+    }
 
     /// Specifies edition of [Edition-Based Redefinition][].
     ///
-    /// [Edition-Based Redefinition]: https://docs.oracle.com/database/122/ADFNS/editions.htm#ADFNS020
-    Edition(String),
+    /// [Edition-Based Redefinition]: https://docs.oracle.com/en/database/oracle/oracle-database/19/adfns/editions.html#GUID-58DE05A0-5DEF-4791-8FA8-F04D11964906
+    pub fn edition<S>(&mut self, edition: S) -> &mut Connector
+    where
+        S: Into<String>,
+    {
+        self.edition = edition.into();
+        self
+    }
 
     /// Sets the driver name displayed in [V$SESSION_CONNECT_INFO.CLIENT_DRIVER][].
     ///
@@ -245,8 +346,91 @@ pub enum ConnParam {
     /// chracters "rust-ora" are displayed when the Oracle server version is
     /// lower than 12.0.1.2.
     ///
-    /// [V$SESSION_CONNECT_INFO.CLIENT_DRIVER]: https://docs.oracle.com/database/122/REFRN/V-SESSION_CONNECT_INFO.htm
-    DriverName(String),
+    /// [V$SESSION_CONNECT_INFO.CLIENT_DRIVER]: https://docs.oracle.com/en/database/oracle/oracle-database/19/refrn/V-SESSION_CONNECT_INFO.html#GUID-9F0DCAEA-A67E-4183-89E7-B1555DC591CE
+    pub fn driver_name<S>(&mut self, driver_name: S) -> &mut Connector
+    where
+        S: Into<String>,
+    {
+        self.driver_name = driver_name.into();
+        self
+    }
+
+    /// Connect an Oracle server using specified parameters
+    pub fn connect(&self) -> Result<Connection> {
+        let ctxt = Context::get()?;
+        let mut common_params = ctxt.common_create_params;
+        let mut conn_params = ctxt.conn_create_params;
+
+        if let Some(ref privilege) = self.privilege {
+            conn_params.authMode |= match privilege {
+                &Privilege::Sysdba => DPI_MODE_AUTH_SYSDBA,
+                &Privilege::Sysoper => DPI_MODE_AUTH_SYSOPER,
+                &Privilege::Sysasm => DPI_MODE_AUTH_SYSASM,
+                &Privilege::Sysbackup => DPI_MODE_AUTH_SYSBKP,
+                &Privilege::Sysdg => DPI_MODE_AUTH_SYSDGD,
+                &Privilege::Syskm => DPI_MODE_AUTH_SYSKMT,
+                &Privilege::Sysrac => DPI_MODE_AUTH_SYSRAC,
+            };
+        }
+        if self.external_auth {
+            conn_params.externalAuth = 1;
+        }
+        if self.prelim_auth {
+            conn_params.authMode |= DPI_MODE_AUTH_PRELIM;
+        }
+        let s = to_odpi_str(&self.new_password);
+        conn_params.newPassword = s.ptr;
+        conn_params.newPasswordLength = s.len;
+        if let Some(purity) = self.purity {
+            conn_params.purity = match purity {
+                Purity::New => DPI_PURITY_NEW,
+                Purity::Self_ => DPI_PURITY_SELF,
+            };
+        }
+        let s = to_odpi_str(&self.connection_class);
+        conn_params.connectionClass = s.ptr;
+        conn_params.connectionClassLength = s.len;
+        let mut app_context = Vec::with_capacity(self.app_context.len());
+        for ac in &self.app_context {
+            let namespace = to_odpi_str(&ac.0);
+            let name = to_odpi_str(&ac.1);
+            let value = to_odpi_str(&ac.2);
+            app_context.push(dpiAppContext {
+                namespaceName: namespace.ptr,
+                namespaceNameLength: namespace.len,
+                name: name.ptr,
+                nameLength: name.len,
+                value: value.ptr,
+                valueLength: value.len,
+            });
+        }
+        if app_context.len() != 0 {
+            conn_params.appContext = app_context.as_mut_ptr();
+            conn_params.numAppContext = app_context.len() as u32;
+        }
+        let s = to_odpi_str(&self.tag);
+        conn_params.tag = s.ptr;
+        conn_params.tagLength = s.len;
+        if self.match_any_tag {
+            conn_params.matchAnyTag = 1;
+        }
+        if self.events {
+            common_params.createMode |= DPI_MODE_CREATE_EVENTS;
+        }
+        let s = to_odpi_str(&self.edition);
+        common_params.edition = s.ptr;
+        common_params.editionLength = s.len;
+        let s = to_odpi_str(&self.driver_name);
+        common_params.driverName = s.ptr;
+        common_params.driverNameLength = s.len;
+        Connection::connect_internal(
+            &self.username,
+            &self.password,
+            &self.connect_string,
+            Some(common_params),
+            Some(conn_params),
+        )
+    }
 }
 
 /// Connection to an Oracle database
@@ -260,14 +444,19 @@ pub struct Connection {
 }
 
 impl Connection {
-    /// Connects to an Oracle server
+    /// Connects to an Oracle server using username, password and connect string.
+    ///
+    /// If you need to connect the server with additional parameters
+    /// such as SYSDBA privilege, use [Connector] instead.
+    ///
+    /// [Connector]: struct.Connector.html
     ///
     /// # Examples
     /// Connect to a local database.
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// let conn = Connection::connect("scott", "tiger", "", &[])?;
+    /// let conn = Connection::connect("scott", "tiger", "")?;
     /// # Ok(())} fn main() { try_main().unwrap(); }
     /// ```
     ///
@@ -276,148 +465,34 @@ impl Connection {
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
     /// let conn = Connection::connect("scott", "tiger",
-    ///                                "server_name:1521/service_name", &[])?;
+    ///                                "server_name:1521/service_name")?;
     /// # Ok(())} fn main() { try_main().unwrap(); }
     /// ```
-    ///
-    /// Connect as sysdba.
-    ///
-    /// ```no_run
-    /// # use oracle::*; fn try_main() -> Result<()> {
-    /// let conn = Connection::connect("system", "manager", "",
-    ///                                &[ConnParam::Sysdba])?;
-    /// # Ok(())} fn main() { try_main().unwrap(); }
-    /// ```
-    ///
-    pub fn connect(
-        username: &str,
-        password: &str,
-        connect_string: &str,
-        params: &[ConnParam],
-    ) -> Result<Connection> {
-        let ctxt = Context::get()?;
+    pub fn connect<U, P, C>(username: U, password: P, connect_string: C) -> Result<Connection>
+    where
+        U: AsRef<str>,
+        P: AsRef<str>,
+        C: AsRef<str>,
+    {
         Connection::connect_internal(
-            ctxt,
-            username,
-            password,
-            connect_string,
-            params,
-            ptr::null_mut(),
+            username.as_ref(),
+            password.as_ref(),
+            connect_string.as_ref(),
+            None,
+            None,
         )
     }
 
-    #[deprecated(since = "0.0.6", note = "use `Connection::connect()` instead")]
-    #[doc(hidden)]
-    pub fn new(username: &str, password: &str, connect_string: &str) -> Result<Connection> {
-        Connection::connect(username, password, connect_string, &[])
-    }
-
     pub(crate) fn connect_internal(
-        ctxt: &'static Context,
         username: &str,
         password: &str,
         connect_string: &str,
-        params: &[ConnParam],
-        pool: *mut dpiPool,
+        common_params: Option<dpiCommonCreateParams>,
+        conn_params: Option<dpiConnCreateParams>,
     ) -> Result<Connection> {
-        let mut common_params = ctxt.common_create_params;
-        let mut conn_params = ctxt.conn_create_params;
-
-        let mut num_app_context = 0;
-        for param in params {
-            if let &ConnParam::AppContext(_, _, _) = param {
-                num_app_context += 1;
-            }
-        }
-        let mut app_context = Vec::with_capacity(num_app_context);
-
-        for param in params {
-            match param {
-                &ConnParam::Sysdba => {
-                    conn_params.authMode |= DPI_MODE_AUTH_SYSDBA;
-                }
-                &ConnParam::Sysoper => {
-                    conn_params.authMode |= DPI_MODE_AUTH_SYSOPER;
-                }
-                &ConnParam::Sysasm => {
-                    conn_params.authMode |= DPI_MODE_AUTH_SYSASM;
-                }
-                &ConnParam::Sysbackup => {
-                    conn_params.authMode |= DPI_MODE_AUTH_SYSBKP;
-                }
-                &ConnParam::Sysdg => {
-                    conn_params.authMode |= DPI_MODE_AUTH_SYSDGD;
-                }
-                &ConnParam::Syskm => {
-                    conn_params.authMode |= DPI_MODE_AUTH_SYSKMT;
-                }
-                &ConnParam::Sysrac => {
-                    conn_params.authMode |= DPI_MODE_AUTH_SYSRAC;
-                }
-                &ConnParam::ExternalAuth => {
-                    conn_params.externalAuth = 1;
-                }
-                &ConnParam::PrelimAuth => {
-                    conn_params.authMode |= DPI_MODE_AUTH_PRELIM;
-                }
-                &ConnParam::NewPassword(ref password) => {
-                    let s = to_odpi_str(password);
-                    conn_params.newPassword = s.ptr;
-                    conn_params.newPasswordLength = s.len;
-                }
-                &ConnParam::PurityNew => {
-                    conn_params.purity = DPI_PURITY_NEW;
-                }
-                &ConnParam::PuritySelf => {
-                    conn_params.purity = DPI_PURITY_SELF;
-                }
-                &ConnParam::ConnectionClass(ref name) => {
-                    let s = to_odpi_str(name);
-                    conn_params.connectionClass = s.ptr;
-                    conn_params.connectionClassLength = s.len;
-                }
-                &ConnParam::AppContext(ref namespace, ref name, ref value) => {
-                    let namespace = to_odpi_str(namespace);
-                    let name = to_odpi_str(name);
-                    let value = to_odpi_str(value);
-                    app_context.push(dpiAppContext {
-                        namespaceName: namespace.ptr,
-                        namespaceNameLength: namespace.len,
-                        name: name.ptr,
-                        nameLength: name.len,
-                        value: value.ptr,
-                        valueLength: value.len,
-                    });
-                }
-                &ConnParam::Tag(ref tag) => {
-                    let s = to_odpi_str(tag);
-                    conn_params.tag = s.ptr;
-                    conn_params.tagLength = s.len;
-                }
-                &ConnParam::MatchAnyTag => {
-                    conn_params.matchAnyTag = 1;
-                }
-                &ConnParam::Events => {
-                    common_params.createMode |= DPI_MODE_CREATE_EVENTS;
-                }
-                &ConnParam::Edition(ref edition) => {
-                    let s = to_odpi_str(edition);
-                    common_params.edition = s.ptr;
-                    common_params.editionLength = s.len;
-                }
-                &ConnParam::DriverName(ref driver_name) => {
-                    let s = to_odpi_str(driver_name);
-                    common_params.driverName = s.ptr;
-                    common_params.driverNameLength = s.len;
-                }
-            }
-        }
-        if app_context.len() != 0 {
-            conn_params.appContext = app_context.as_mut_ptr();
-            conn_params.numAppContext = app_context.len() as u32;
-        }
-        conn_params.pool = pool;
-
+        let ctxt = Context::get()?;
+        let common_params = common_params.unwrap_or(ctxt.common_create_params);
+        let mut conn_params = conn_params.unwrap_or(ctxt.conn_create_params);
         let username = to_odpi_str(username);
         let password = to_odpi_str(password);
         let connect_string = to_odpi_str(connect_string);
@@ -462,7 +537,7 @@ impl Connection {
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// # let conn = Connection::connect("scott", "tiger", "", &[])?;
+    /// # let conn = Connection::connect("scott", "tiger", "")?;
     /// let mut stmt = conn.prepare("insert into emp(empno, ename) values (:id, :name)", &[])?;
     ///
     /// let emp_list = [
@@ -496,7 +571,7 @@ impl Connection {
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// # let conn = Connection::connect("scott", "tiger", "", &[])?;
+    /// # let conn = Connection::connect("scott", "tiger", "")?;
     /// // fetch top 10 rows.
     /// let mut stmt = conn.prepare("select * from (select empno, ename from emp order by empno) where rownum <= 10",
     ///                             &[StmtParam::FetchArraySize(10)])?;
@@ -636,7 +711,7 @@ impl Connection {
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// let conn = Connection::connect("scott", "tiger", "", &[])?;
+    /// let conn = Connection::connect("scott", "tiger", "")?;
     ///
     /// // execute a statement without bind parameters
     /// conn.execute("insert into emp(empno, ename) values (113, 'John')", &[])?;
@@ -661,7 +736,7 @@ impl Connection {
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// let conn = Connection::connect("scott", "tiger", "", &[])?;
+    /// let conn = Connection::connect("scott", "tiger", "")?;
     ///
     /// // execute a statement with binding parameters by name
     /// conn.execute_named("insert into emp(empno, ename) values (:id, :name)",
@@ -710,7 +785,7 @@ impl Connection {
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// let conn = Connection::connect("scott", "tiger", "", &[])?;
+    /// let conn = Connection::connect("scott", "tiger", "")?;
     /// let objtype = conn.object_type("MDSYS.SDO_GEOMETRY");
     /// # Ok(())} fn main() { try_main().unwrap(); }
     /// ```
@@ -769,7 +844,7 @@ impl Connection {
     ///
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
-    /// let conn = Connection::connect("scott", "tiger", "", &[])?;
+    /// let conn = Connection::connect("scott", "tiger", "")?;
     /// let (version, banner) = conn.server_version()?;
     /// println!("Oracle Version: {}", version);
     /// println!("--- Version Banner ---");
@@ -907,7 +982,7 @@ impl Connection {
     /// without executing a statement. The module name is piggybacked
     /// to the server with the next network round-trip.
     ///
-    /// [DBMS_APPLICATION_INFO.SET_MODULE]: https://docs.oracle.com/database/122/ARPLS/DBMS_APPLICATION_INFO.htm#GUID-B2E2BD20-D91D-40DB-A3F6-37A853384F30
+    /// [DBMS_APPLICATION_INFO.SET_MODULE]: https://docs.oracle.com/en/database/oracle/oracle-database/19/arpls/DBMS_APPLICATION_INFO.html#GUID-B2E2BD20-D91D-40DB-A3F6-37A853384F30
     pub fn set_module(&self, module: &str) -> Result<()> {
         let s = to_odpi_str(module);
         chkerr!(self.ctxt, dpiConn_setModule(self.handle, s.ptr, s.len));
@@ -920,7 +995,7 @@ impl Connection {
     /// without executing a statement. The action name is piggybacked
     /// to the server with the next network round-trip.
     ///
-    /// [DBMS_APPLICATION_INFO.SET_ACTION]: https://docs.oracle.com/database/122/ARPLS/DBMS_APPLICATION_INFO.htm#GUID-90DA860F-BFBE-4539-BA00-2279B02B8F26
+    /// [DBMS_APPLICATION_INFO.SET_ACTION]: https://docs.oracle.com/en/database/oracle/oracle-database/19/arpls/DBMS_APPLICATION_INFO.html#GUID-90DA860F-BFBE-4539-BA00-2279B02B8F26
     pub fn set_action(&self, action: &str) -> Result<()> {
         let s = to_odpi_str(action);
         chkerr!(self.ctxt, dpiConn_setAction(self.handle, s.ptr, s.len));
@@ -933,7 +1008,7 @@ impl Connection {
     /// without executing a statement. The client info is piggybacked
     /// to the server with the next network round-trip.
     ///
-    /// [DBMS_APPLICATION_INFO.SET_CLIENT_INFO]: https://docs.oracle.com/database/122/ARPLS/DBMS_APPLICATION_INFO.htm#GUID-68A3DF04-BE91-46CC-8D2B-97BA0E89956F
+    /// [DBMS_APPLICATION_INFO.SET_CLIENT_INFO]: https://docs.oracle.com/en/database/oracle/oracle-database/19/arpls/DBMS_APPLICATION_INFO.html#GUID-68A3DF04-BE91-46CC-8D2B-97BA0E89956F
     pub fn set_client_info(&self, client_info: &str) -> Result<()> {
         let s = to_odpi_str(client_info);
         chkerr!(self.ctxt, dpiConn_setClientInfo(self.handle, s.ptr, s.len));
@@ -946,7 +1021,7 @@ impl Connection {
     /// without executing a statement. The client identifier is piggybacked
     /// to the server with the next network round-trip.
     ///
-    /// [DBMS_SESSION.SET_IDENTIFIER]: https://docs.oracle.com/database/122/ARPLS/DBMS_SESSION.htm#GUID-988EA930-BDFE-4205-A806-E54F05333562
+    /// [DBMS_SESSION.SET_IDENTIFIER]: https://docs.oracle.com/en/database/oracle/oracle-database/19/arpls/DBMS_SESSION.html#GUID-988EA930-BDFE-4205-A806-E54F05333562
     pub fn set_client_identifier(&self, client_identifier: &str) -> Result<()> {
         let s = to_odpi_str(client_identifier);
         chkerr!(
@@ -965,9 +1040,9 @@ impl Connection {
     ///
     /// See [Monitoring Database Operations][] in Oracle Database SQL Tuning Guide
     ///
-    /// [db_op]: https://docs.oracle.com/database/121/TGSQL/glossary.htm#GUID-EB7D5D0A-0439-4336-8DC3-2DA24072977F
-    /// [DBMS_SQL_MONITOR.BEGIN_OPERATION]: https://docs.oracle.com/database/122/ARPLS/DBMS_SQL_MONITOR.htm#ARPLS74785
-    /// [Monitoring Database Operations]: https://docs.oracle.com/database/122/TGSQL/monitoring-database-operations.htm#TGSQL-GUID-C941CE9D-97E1-42F8-91ED-4949B2B710BF
+    /// [db_op]: https://docs.oracle.com/en/database/oracle/oracle-database/19/tgsql/glossary.html#GUID-9CE3C342-D210-4690-A7E9-5813EF9D558E
+    /// [DBMS_SQL_MONITOR.BEGIN_OPERATION]: https://docs.oracle.com/en/database/oracle/oracle-database/19/arpls/DBMS_SQL_MONITOR.html#GUID-25BE0E79-3A19-4303-9F66-2CFDB87C7F82
+    /// [Monitoring Database Operations]: https://docs.oracle.com/en/database/oracle/oracle-database/19/tgsql/monitoring-database-operations.html#GUID-C941CE9D-97E1-42F8-91ED-4949B2B710BF
     pub fn set_db_op(&self, db_op: &str) -> Result<()> {
         let s = to_odpi_str(db_op);
         chkerr!(self.ctxt, dpiConn_setDbOp(self.handle, s.ptr, s.len));
@@ -990,19 +1065,19 @@ impl Connection {
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
     /// // connect as sysdba with prelim_auth mode
-    /// let conn = Connection::connect("sys", "change_on_install", "",
-    ///                                 &[ConnParam::Sysdba,
-    ///                                   ConnParam::PrelimAuth,
-    ///                                   ])?;
+    /// let conn = Connector::new("sys", "change_on_install", "")
+    ///     .privilege(Privilege::Sysdba)
+    ///     .prelim_auth(true)
+    ///     .connect()?;
     ///
     /// // start the instance
     /// conn.startup_database(&[])?;
     /// conn.close()?;
     ///
     /// // connect again without prelim_auth
-    /// let conn = Connection::connect("sys", "change_on_install", "",
-    ///                                 &[ConnParam::Sysdba,
-    ///                                   ])?;
+    /// let conn = Connector::new("sys", "change_on_install", "")
+    ///     .privilege(Privilege::Sysdba)
+    ///     .connect()?;
     ///
     /// // mount and open a database
     /// conn.execute("alter database mount", &[])?;
@@ -1063,8 +1138,9 @@ impl Connection {
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
     /// // connect as sysdba
-    /// let conn = Connection::connect("sys", "change_on_install", "",
-    ///                                &[ConnParam::Sysdba])?;
+    /// let conn = Connector::new("sys", "change_on_install", "")
+    ///     .privilege(Privilege::Sysdba)
+    ///     .connect()?;
     ///
     /// // begin 'shutdown immediate'
     /// conn.shutdown_database(ShutdownMode::Immediate)?;
@@ -1083,8 +1159,8 @@ impl Connection {
     /// ```no_run
     /// # use oracle::*; fn try_main() -> Result<()> {
     /// // connect as sysdba
-    /// let conn = Connection::connect("sys", "change_on_install", "",
-    ///                                &[ConnParam::Sysdba])?;
+    /// let conn = Connector::new("sys", "change_on_install", "")
+    ///     .privilege(Privilege::Sysdba).connect()?;
     ///
     /// // 'shutdown abort'
     /// conn.shutdown_database(ShutdownMode::Abort)?;