+++ src/sql_type/timestamp.rs
@@ -73,7 +73,7 @@ use ParseOracleTypeError;
 ///
 /// ```no_run
 /// # use oracle::*; use oracle::sql_type::*; fn try_main() -> Result<()> {
-/// let conn = Connection::connect("scott", "tiger", "", &[])?;
+/// let conn = Connection::connect("scott", "tiger", "")?;
 ///
 /// // Fetch Timestamp
 /// let sql = "select TIMESTAMP '2017-08-09 11:22:33.500' from dual";